chore: Move footer model to inside footer directory
