+++ src/components/organisms/footer/type/model.ts
