+++ src/models/index.ts
@@ -1,6 +1,5 @@
 export * from "./contact-section";
 export * from "./festa-section";
-export * from "./footer";
 export * from "./hero-section";
 export * from "./layout";
 export * from "./map-section";
